+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -173,6 +173,7 @@ export const ScreenName = {
   CosmosDelegationConnectDevice: "CosmosDelegationConnectDevice",
   CosmosDelegationValidationError: "CosmosDelegationValidationError",
   CosmosDelegationValidationSuccess: "CosmosDelegationValidationSuccess",
+  CosmosDefaultRedelegationAmount: "CosmosDefaultRedelegationAmount",
   CosmosRedelegationValidator: "CosmosRedelegationValidator",
   CosmosRedelegationAmount: "CosmosRedelegationAmount",
   CosmosRedelegationSelectDevice: "CosmosRedelegationSelectDevice",
@@ -384,6 +385,11 @@ export const ScreenName = {
   CardanoEditMemo: "CardanoEditMemo",
   // hedera
   HederaEditMemo: "HederaEditMemo",
+
+  CustomImageStep1Crop: "CustomImageStep1Crop",
+  CustomImageStep2Preview: "CustomImageStep2Preview",
+  CustomImageStep3Transfer: "CustomImageStep3Transfer",
+  CustomImageErrorScreen: "CustomImageErrorScreen",
 };
 export const NavigatorName = {
   // Stack
@@ -461,4 +467,9 @@ export const NavigatorName = {
   // Root
   RootNavigator: "RootNavigator",
   Discover: "Discover",
+
+  // Custom Image
+  CustomImage: "CustomImage",
+
+  WalletConnect: "WalletConnect",
 };